refactor(schema-api): extract DatabaseApi trait from SchemaApi
Extract database operations into a separate trait following single
responsibility principle. This creates cleaner separation between
database and table operations.

- Create DatabaseApi trait with 9 database operations
- Add database_util module for database-specific utilities
